+++ tiled/client/context.py
@@ -6,7 +6,7 @@ import time
 import urllib.parse
 import warnings
 from pathlib import Path
-from typing import List, Literal
+from typing import List, Literal, Optional
 from urllib.parse import parse_qs, urlparse
 
 import httpx
@@ -103,6 +103,9 @@ def prompt_for_credentials(http_client, providers: List[AboutAuthenticationProvi
         spec = identity_provider_input(providers)
     auth_endpoint = spec.links["auth_endpoint"]
     provider = spec.provider
+    client_id = spec.links.get("client_id")
+    token_endpoint = spec.links.get("token_endpoint")
+    oauth2_spec = True if client_id and token_endpoint else False
     mode = spec.mode
     # Note: "password" is included here for back-compat with older servers;
     # the new name for this mode is "internal".
@@ -134,12 +137,14 @@ def prompt_for_credentials(http_client, providers: List[AboutAuthenticationProvi
     elif mode == "external":
         # Display link and access code, and try to open web browser.
         # Block while polling the server awaiting confirmation of authorization.
-        tokens = device_code_grant(http_client, auth_endpoint)
+        tokens = device_code_grant(
+            http_client, auth_endpoint, client_id, token_endpoint
+        )
     else:
         raise ValueError(f"Server has unknown authentication mechanism {mode!r}")
     confirmation_message = spec.confirmation_message
     if confirmation_message:
-        username = tokens["identity"]["id"]
+        username = "jwt_token" if oauth2_spec else tokens["identity"]["id"]
         print(confirmation_message.format(id=username))
     return tokens
 
@@ -1010,13 +1015,26 @@ def password_grant(http_client, auth_endpoint, provider, username, password):
     return token_response.json()
 
 
-def device_code_grant(http_client, auth_endpoint):
+def device_code_grant(
+    http_client,
+    auth_endpoint: str,
+    client_id: Optional[str],
+    token_endpoint: Optional[str],
+):
+    if client_id and token_endpoint:
+        oauth2_spec = True
+        data = {"client_id": client_id}
+        uri = "verification_uri_complete"
+    else:
+        oauth2_spec = False
+        data = {}
+        uri = "authorization_uri"
     for attempt in retry_context():
         with attempt:
-            verification_response = http_client.post(auth_endpoint, json={}, auth=None)
+            verification_response = http_client.post(auth_endpoint, data=data)
             handle_error(verification_response)
     verification = verification_response.json()
-    authorization_uri = verification["authorization_uri"]
+    authorization_uri = verification[uri]
     print(
         f"""
 You have {int(verification['expires_in']) // 60} minutes to visit this URL
@@ -1042,19 +1060,34 @@ and enter the code:
             with attempt:
                 # Intentionally do not wrap this in handle_error(...).
                 # Check status codes manually below.
-                access_response = http_client.post(
-                    verification["verification_uri"],
-                    json={
-                        "device_code": verification["device_code"],
-                        "grant_type": "urn:ietf:params:oauth:grant-type:device_code",
-                    },
-                    auth=None,
-                )
-                if (access_response.status_code == httpx.codes.BAD_REQUEST) and (
-                    access_response.json()["detail"]["error"] == "authorization_pending"
-                ):
-                    print(".", end="", flush=True)
-                    continue
+                if oauth2_spec:
+                    access_response = http_client.post(
+                        token_endpoint,
+                        data={
+                            "device_code": verification["device_code"],
+                            "grant_type": "urn:ietf:params:oauth:grant-type:device_code",
+                            "client_id": client_id,
+                        },
+                        headers={"Content-Type": "application/x-www-form-urlencoded"},
+                    )
+                else:
+                    access_response = http_client.post(
+                        verification["verification_uri"],
+                        json={
+                            "device_code": verification["device_code"],
+                            "grant_type": "urn:ietf:params:oauth:grant-type:device_code",
+                        },
+                        auth=None,
+                    )
+                if access_response.status_code == httpx.codes.BAD_REQUEST:
+                    access_response_error = (
+                        access_response.json()["error"]
+                        if oauth2_spec
+                        else access_response.json()["detail"]["error"]
+                    )
+                    if access_response_error == "authorization_pending":
+                        print(".", end="", flush=True)
+                        continue
                 handle_error(access_response)
         print("")
         break